Completely isolate schema definitions and dependencies
Allows for a more "pluggable" schema architecture
Update README.md
