+++ bindings/CXX17/ncio.h
@@ -8,8 +8,12 @@
 
 #pragma once
 
-#include "ncio/common/ncioMacros.h"
 #include "ncio/common/ncioTypes.h"
 #include "ncio/cxx17/cxx17DataDescriptor.h"
 #include "ncio/cxx17/cxx17NCIO.h"
-#include "ncio/cxx17/schema/cxx17ncioNexus.h"
+#include "ncio/ncioConfig.h"
+
+#ifdef NCIO_HAVE_SCHEMA_NEXUS
+#include "ncio/cxx17/schema/nexus/cxx17ncioNexus.h"
+#include "ncio/schema/nexus/ncioTypesNexus.h"
+#endif
