+++ bindings/CXX17/ncio/cxx17/schema/nexus/cxx17ncioNexus.h
@@ -1,6 +1,6 @@
 /**
- * cxx17nexus.h : public interface for functions inside the ncio::nexus
- * workspace targeting the Nexus schema
+ * cxx17ncioNexus.h : public interface for functions inside the
+ * ncio::schema::nexus workspace targeting the Nexus schema
  *
  *  Created on: May 26, 2020
  *      Author: William F Godoy godoywf@ornl.gov
@@ -8,11 +8,9 @@
 
 #pragma once
 
-#include "ncio/common/ncioTypesNexus.h"
-
 #include <string>
 
-namespace ncio::nexus
+namespace ncio::schema::nexus
 {
 
 /**
