+++ source/ncio/core/DataDescriptor.tcc
@@ -10,42 +10,11 @@
 
 #include "DataDescriptor.h"
 
-#include "ncio/common/ncioTypes.h"
-#include "ncio/schema/ncioNexus.h"
+#include "ncio/ncioConfig.h"
 
 namespace ncio::core
-{
+{} // end namespace ncio::core
 
-#define NCIO_CORE_DATADESCRIPTOR_PUTGET_NEXUS(bank, entry, T)                  \
-    template <>                                                                \
-    void DataDescriptor::Put<bank, entry>(const T *data)                       \
-    {                                                                          \
-        m_Transport->Put(nexus::impl::ToString<bank, entry>(), data);          \
-    }                                                                          \
-                                                                               \
-    template <>                                                                \
-    void DataDescriptor::Get<bank, entry>(T * data)                            \
-    {                                                                          \
-        m_Transport->Get(nexus::impl::ToString<bank, entry>(), data);          \
-    }
-
-#define declare_ncio_nexus_bank_entries(T)                                     \
-    NCIO_CORE_DATADESCRIPTOR_PUTGET_NEXUS(                                     \
-        nexus::bank##T, nexus::bank##T::event_id, std::uint64_t)               \
-                                                                               \
-    NCIO_CORE_DATADESCRIPTOR_PUTGET_NEXUS(                                     \
-        nexus::bank##T, nexus::bank##T::event_index, std::uint32_t)            \
-                                                                               \
-    NCIO_CORE_DATADESCRIPTOR_PUTGET_NEXUS(                                     \
-        nexus::bank##T, nexus::bank##T::event_time_offset, float)              \
-                                                                               \
-    NCIO_CORE_DATADESCRIPTOR_PUTGET_NEXUS(                                     \
-        nexus::bank##T, nexus::bank##T::event_time_zero, double)               \
-                                                                               \
-    NCIO_CORE_DATADESCRIPTOR_PUTGET_NEXUS(nexus::bank##T,                      \
-                                          nexus::bank##T::total_counts, float)
-
-NCIO_MACRO_NEXUS_FOREACH_BANK_ID(declare_ncio_nexus_bank_entries)
-#undef declare_ncio_nexus_bank_entries
-
-} // end namespace ncio::core
+#ifdef NCIO_HAVE_SCHEMA_NEXUS
+#include "ncio/schema/nexus/ncioNexusDataDescriptor.tcc"
+#endif
