+++ testing/functional/bindings/CXX17/functionalTest_cxx17NCIO.cpp
@@ -22,7 +22,7 @@ TEST_CASE("Function test for ncio::NCIO C++17 bindings class")
         ncio::DataDescriptor fh = ncio.Open("dummy", ncio::openmode::read);
 
         float total_counts = 0;
-        fh.Get<ncio::nexus::bank1::total_counts>(&total_counts);
+        fh.Get<ncio::schema::nexus::bank1::total_counts>(&total_counts);
         fh.Execute();
 
         CHECK(total_counts == 0.f);
